+++ lib/interfaces/IAssertion.ts
@@ -16,6 +16,7 @@ interface IEqualityAssertion {
   toStrictEqual(expected: any): IAssertion;
   toStrictNotEqual(expected: any): IAssertion;
   toBe(expected: any): IAssertion;
+  toNotBe(expected: any): IAssertion;
 }
 
 interface ITruthinessAssertion {