+++ lib/interfaces/IStestOptions.ts
@@ -1,0 +1,8 @@
+import { OptionsType } from '@stlib/utils';
+import { InitValue } from '../../bin/cli';
+
+export interface StestOptions extends Partial<OptionsType> {
+  config?: string;
+  init?: InitValue;
+  watch?: boolean;
+}