+++ lib/interfaces/index.ts
@@ -2,3 +2,4 @@ export * from './ITestSuite';
 export * from './ITestCase';
 export * from './IAssertion';
 export * from './IAfter_Before';
+export * from './IStestOptions';