+++ lib/types/config.type.ts
@@ -1,0 +1,6 @@
+export type ConfigType = {
+  pattern?: string;
+  ignore?: string[];
+  cacheWatcher?: boolean;
+  autoClearMocks?: boolean;
+};