+++ lib/types/watcher.events.ts
@@ -1,0 +1,11 @@
+export type ActionSymbolType = {
+  add: string;
+  unlink: string;
+  change: string;
+};
+
+export const actionSymbol: ActionSymbolType = {
+  add: '+',
+  unlink: '-',
+  change: '↻',
+};