+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Unreleased
+### Added:
+* Now test cases can be asynchronous
+
 ### Patch 1.0.2
 ### Added:
 * Caching option in config for watch mode. Now it can be disabled