+++ CHANGELOG.md
@@ -5,11 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## Patch 1.0.4
 ### Added:
 * Data decorators for tests
 * Optimized mocks registry and test running
 * Spy functionality
+* Timeout for tests
+
+### Changed:
+* Optimized test registry
+* Optimized mock registry
 
 ## Patch 1.0.3
 ### Added: